Removes duplication of dates in the post front matter.
The post date will be taken from the filename and does not need to be
repeated in the front matter. Leaving it in the front matter causes
confusion and mistakes when the post is copied between blogs as the
committer can forget to change the date in both places, causing broken
images.